remove environment-context specific when importing symbol in the import engine and use environement's level instead
